# docs/yapi-ptz-biz-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "视频中台 - 云台监控 / 业务端",
    "description": "仅含云台监控与业务端（目录、设备、直播、录像、录像机回放、片段截取）。\n除目录树、设备详情、录像机回放、片段文件外，带参接口一律 POST + JSON Body。\n统一响应 { code, message, data }；code=0 成功。时间字段为毫秒时间戳。\n设备 status：0=不可用 1=已启用 2=已停用。直播仅 status=1；回放/截取 status=2 不可用。\n一台设备只对接一个平台 vendor：MOCK 模拟、UNIVIEW 宇视、HIKVISION 海康。调用方不传平台，只传 deviceId，服务按 vendor 调用对应平台。云台路径为 /api/ptz。海康尚未实现，失败文案为「海康尚未对接」。\n返回只有调用方会用到的字段。没有密码、经纬度、创建时间、服务器磁盘路径。\n宇视云台发给这台摄像机自己的地址；历史录像在它绑定的录像机上。示例设备 UV_10135。\n鉴权：auth.enabled=false 时免登录；true 时 Header Authorization: Bearer {token}。<video> 带不了 Header，URL 上加 token。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://8.130.74.232:8090",
      "description": "线上环境"
    },
    {
      "url": "http://127.0.0.1:8090",
      "description": "本地环境"
    }
  ],
...
  "paths": {
...
    "/api/ptz/devices/{deviceId}": {
      "get": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-设备详情（含码流，不拉流）",
        "description": "取这台设备已启用码流的播放地址。不拉流。宇视读这台摄像机的码流；模拟读中台已登记的码流。海康失败：海康尚未对接。未配置宇视地址时失败：设备未配置宇视地址。要出画面再调 /api/preview/start。",
        "operationId": "univiewDevice",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "description": "设备编码",
            "schema": {
              "type": "string",
              "description": "设备编码",
              "example": "UV_10135"
            },
            "example": "UV_10135"
          }
        ],
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "统一 JSON 响应：code / message / data",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "description": "提示信息：成功一般为 success，失败为错误原因",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deviceId": {
                          "type": "string",
                          "description": "设备编码",
                          "example": "UV_10135"
                        },
                        "name": {
                          "type": "string",
                          "description": "名称",
                          "example": "演示球机"
                        },
                        "streams": {
                          "type": "array",
                          "description": "已启用码流",
                          "items": {
                            "$ref": "#/components/schemas/CatalogStream"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "code": 0,
                    "message": "success",
                    "data": "云台设备详情"
                  }
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "deviceId": "UV_10135",
                    "name": "演示球机",
                    "streams": [
                      {
                        "channelType": "visible",
                        "streamType": "sub",
                        "streamUrl": "http://8.130.74.232:8080/live/uv_UV_10135_sub.live.flv"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ptz/move": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-方向移动",
        "description": "连续转动，按住时重复调用，松开传 direction=stop。宇视发到这台摄像机。模拟只返回成功，不转动设备。海康失败：海康尚未对接。\ndirection：up/down/left/right/left_up/left_down/right_up/right_down/stop。其它值返回 400。\n使用该设备自己的 IP、端口和账号，不再使用配置文件里的单一摄像机。",
        "operationId": "ptzMove",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzMoveRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "direction": "up",
                "speed": 4
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "move",
                    "deviceId": "UV_10135",
                    "ptzCmd": 1026,
                    "speed": 4,
                    "response": "{\"Response\":{\"ResponseCode\":0}}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ptz/zoom": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-变倍",
        "description": "点按一次：下发变倍后短暂保持再停止。宇视发到这台摄像机。模拟只返回成功。海康失败：海康尚未对接。action：in / zoom_in / tele 拉近；out / zoom_out / wide 拉远。其它值返回 400。",
        "operationId": "ptzZoom",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzActionRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "action": "zoom_in",
                "speed": 4
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "zoom",
                    "deviceId": "UV_10135",
                    "speed": 4
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ptz/focus": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-对焦",
        "description": "点按一次：下发对焦后短暂保持再停止。宇视发到这台摄像机。模拟只返回成功。海康失败：海康尚未对接。action：near / focus_near 近焦；far / focus_far 远焦。其它值返回 400。",
        "operationId": "ptzFocus",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzActionRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "action": "focus_near",
                "speed": 4
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "focus",
                    "deviceId": "UV_10135",
                    "speed": 4
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ptz/wide-angle": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-一键广角",
        "description": "点按一次拉远（广角）。宇视发到这台摄像机。模拟只返回成功。海康失败：海康尚未对接。速度固定 4。",
        "operationId": "ptzWideAngle",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzDeviceIdRequest"
              },
              "example": {
                "deviceId": "UV_10135"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "wide-angle",
                    "deviceId": "UV_10135",
                    "speed": 4
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ptz/preset/goto": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-调用预置位",
        "description": "转到已保存预置位。宇视发到这台摄像机。模拟只确认中台预置位存在。海康失败：海康尚未对接。编号与保存时的 index 一致。",
        "operationId": "ptzGotoPreset",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzPresetGotoRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "index": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "goto-preset",
                    "deviceId": "UV_10135",
                    "presetIndex": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ptz/preset/save": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-保存预置位",
        "description": "把当前姿态存为预置位。宇视写到这台摄像机。模拟只写入中台。海康失败：海康尚未对接。name 不能为空。已存在且 overwrite=false 时失败：预置位 n 已存在，请勾选覆盖。",
        "operationId": "ptzSetPreset",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzPresetSaveRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "index": 1,
                "name": "东门全景",
                "overwrite": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "set-preset",
                    "deviceId": "UV_10135",
                    "presetIndex": 1,
                    "overwrite": false,
                    "name": "门口"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ptz/snapshot": {
      "post": {
        "tags": [
          "云台监控"
        ],
        "summary": "云台-抓拍",
        "description": "宇视向这台摄像机抓拍，结果在 response 字符串。模拟返回占位结果。海康失败：海康尚未对接。channelType 默认 visible。",
        "operationId": "ptzSnapshot",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtzSnapshotRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "channelType": "visible"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespPtz"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "success": true,
                    "mock": false,
                    "live": true,
                    "action": "snapshot",
                    "deviceId": "UV_10135",
                    "channelType": "visible"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogStream": {
        "type": "object",
        "properties": {
          "channelType": {
            "type": "string",
            "description": "可见光为 visible",
            "example": "visible"
          },
          "streamType": {
            "type": "string",
            "description": "main 主码流 / sub 辅码流 / third 第三流",
            "example": "sub"
          },
          "streamUrl": {
            "type": "string",
            "description": "HTTP-FLV。此时 ZLM 上可能还没有流",
            "example": "http://8.130.74.232:8080/live/uv_UV_10135_sub.live.flv"
          }
        },
        "description": "云台设备详情里的已启用码流"
      },
...
      "PtzResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "指令是否已受理",
            "example": true
          },
          "mock": {
            "type": "boolean",
            "description": "是否模拟模式；live 时为 false 或不出现",
            "example": true
          },
          "live": {
            "type": "boolean",
            "description": "是否真实宇视；mock 时不出现",
            "nullable": true
          },
          "action": {
            "type": "string",
            "description": "动作标识：move / zoom / focus / wide-angle / goto-preset / set-preset / snapshot",
            "example": "move"
          },
          "deviceId": {
            "type": "string",
            "description": "设备编码",
            "example": "UV_10135"
          },
          "ptzCmd": {
            "type": "integer",
            "description": "方向/变倍/对焦下发给宇视的 PTZCmd 数值",
            "example": 1026,
            "nullable": true
          },
          "speed": {
            "type": "integer",
            "description": "实际速度，live 会夹到 1~8",
            "example": 4,
            "nullable": true
          },
          "presetIndex": {
            "type": "integer",
            "description": "预置位编号（调用或保存预置位）",
            "example": 1,
            "nullable": true
          },
          "overwrite": {
            "type": "boolean",
            "description": "保存预置位时是否覆盖",
            "example": false,
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "保存后的预置位名称",
            "example": "门口",
            "nullable": true
          },
          "channelType": {
            "type": "string",
            "description": "抓拍通道类型，默认 visible",
            "example": "visible",
            "nullable": true
          },
          "response": {
            "type": "string",
            "description": "宇视 LAPI 原始 JSON 字符串",
            "example": "{\"Response\":{\"ResponseCode\":0}}",
            "nullable": true
          }
        },
        "description": "云台控制结果。live=true 表示已发到该设备自己的摄像机"
      },
...
      "RespPtz": {
        "type": "object",
        "description": "统一 JSON 响应：code / message / data",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息：成功一般为 success，失败为错误原因",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/PtzResult"
          }
        },
        "example": {
          "code": 0,
          "message": "success",
          "data": {
            "success": true,
            "mock": false,
            "live": true,
            "action": "move",
            "deviceId": "UV_10135",
            "ptzCmd": 1026,
            "speed": 4,
            "response": "{\"Response\":{\"ResponseCode\":0}}"
          }
        }
      },
      "PtzMoveRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】云台设备编码",
            "example": "UV_10135"
          },
          "direction": {
            "type": "string",
            "description": "【必填】方向：up / down / left / right / left_up / left_down / right_up / right_down / stop",
            "example": "up",
            "enum": [
              "up",
              "down",
              "left",
              "right",
              "left_up",
              "left_down",
              "right_up",
              "right_down",
              "stop"
            ]
          },
          "speed": {
            "type": "integer",
            "description": "速度，默认 4；live 限制 1~8",
            "example": 4
          }
        },
        "required": [
          "deviceId",
          "direction"
        ],
        "description": "云台方向移动请求"
      },
      "PtzActionRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】云台设备编码",
            "example": "UV_10135"
          },
          "action": {
            "type": "string",
            "description": "【必填】动作。变倍推荐 zoom_in/zoom_out；对焦推荐 focus_near/focus_far",
            "example": "zoom_in"
          },
          "speed": {
            "type": "integer",
            "description": "速度，默认 4",
            "example": 4
          }
        },
        "required": [
          "deviceId",
          "action"
        ],
        "description": "云台变倍/对焦请求"
      },
      "PtzDeviceIdRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】云台设备编码",
            "example": "UV_10135"
          }
        },
        "required": [
          "deviceId"
        ],
        "description": "仅设备编码"
      },
      "PtzPresetGotoRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】云台设备编码",
            "example": "UV_10135"
          },
          "index": {
            "type": "integer",
            "description": "【必填】预置位编号，建议 1~1024",
            "example": 1
          }
        },
        "required": [
          "deviceId",
          "index"
        ],
        "description": "调用预置位请求"
      },
      "PtzPresetSaveRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】云台设备编码",
            "example": "UV_10135"
          },
          "index": {
            "type": "integer",
            "description": "【必填】预置位编号",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "【必填】预置位名称，不能为空或纯空格",
            "example": "东门全景"
          },
          "overwrite": {
            "type": "boolean",
            "description": "是否覆盖已有同号预置位，默认 false",
            "example": false
          }
        },
        "required": [
          "deviceId",
          "index",
          "name"
        ],
        "description": "保存预置位请求"
      },
      "PtzSnapshotRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】云台设备编码",
            "example": "UV_10135"
          },
          "channelType": {
            "type": "string",
            "description": "通道类型：visible=可见光（默认），thermal=热成像",
            "example": "visible"
          }
        },
        "required": [
          "deviceId"
        ],
        "description": "抓拍请求"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "auth.enabled=true 时需要。登录后填 JWT，不要带 Bearer 前缀（YApi 会加）。"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}